Air line almost
I need to fix that jump in the first call but you can see that it ticks
properly. I need to add an array for the lines and then delete them
after the next one is drawn. Does anyone have any idea for a way to use
totalAir to determine the draw for that first one (should only move like
2 degrees like it dos at the end).